qwen4_exp: PR CI green — 500-LoC splits + non-cuda gating for the PLE row cache
CI first saw the full branch after the avarok sync and flagged two jobs:

500-LoC cap (7 violators, all this branch's growth):
  qsa.rs 757          -> prefill_select to qsa_select.rs (#[path] child so
                         private fields stay reachable)  495
  qsa_tests.rs 632    -> prefill-side tests to qsa_tests_prefill.rs (child
                         module; fixture helpers stay super::*)  266
  qwen4_exp.rs 597    -> QSA/PLE/mHC attach + final-norm placeholder to
                         qwen4_exp/aux.rs  477
  qwen3_ssm/init.rs   -> new_sequential to init_sequential.rs  500
  ple/layer.rs        -> Marconi aux-state to ple/aux_state.rs  489
  qwen3_ssm/mod.rs    -> kernel-selection helpers to kernel_select.rs  403
  atlas-kernels lib.rs-> doc trim  500

The split hit the exact trap the process memory warns about: the new mod
decl landed between #[cfg(test)] and the tests decl it guarded, stealing
the attribute — only clippy --all-targets caught the fallout. All four
moved QSA GPU tests re-run PASS on device.

metal (macOS): spark-storage's NgramRowCache is cuda-gated (pinned
GPU-addressable arena) but qwen4_exp/ple.rs consumed it ungated. Same
cuda/non-cuda pair longcat/ngram.rs already carries: the non-cuda load
REFUSES a PLE checkpoint rather than silently serving without the n-gram
injection.

Gates: clippy --all-targets clean, 2306+ tests pass, 4 QSA GPU tests pass,
local cap-check replica reports 0 violators.
