+++ crates/atlas-kernels/src/lib.rs
@@ -252,9 +252,8 @@ pub struct ModelBehavior {
     /// See build_parse_behavior.rs: honor a mid-`<think>` EOS by implicitly
     /// closing the block. Defaults FALSE (pre-p350 behaviour).
     pub honor_eos_inside_thinking: bool,
-    /// A4 POST_THINK_MIN_REASONING floor: suppress `</think>` until this
-    /// many thinking tokens. 16 = historical constant; 0 disables (card-
-    /// native brief-thinking models).
+    /// A4 floor: suppress `</think>` until this many think tokens
+    /// (16 = historical constant; 0 disables — card-native brief thinking).
     pub min_reasoning_floor_tokens: u32,
     /// Cap the thinking budget at 90% of the request's `max_tokens` (true), or
     /// let `max_thinking_budget` be the sole cap (false = vLLM single-budget:
