+++ crates/spark-model/src/layers/ple/layer.rs
@@ -185,52 +185,11 @@ impl PleLayer {
         Ok(())
     }
 
-    /// Marconi aux blob: `[hist_len u32][history u32s][conv f32 bytes]`.
-    /// The whole per-sequence carry — a prefix hit restoring KV+SSM without
-    /// this would run the n-gram hash on the PREVIOUS request's history.
-    pub fn snapshot_aux(&self, gpu: &dyn GpuBackend, stream: u64) -> Result<Vec<u8>> {
-        let st = self
-            .state
-            .lock()
-            .map_err(|_| anyhow::anyhow!("PLE state mutex poisoned"))?;
-        let conv_bytes = self.state_len * self.hc_mult * self.hidden * 4;
-        let mut blob = Vec::with_capacity(4 + st.history.len() * 4 + conv_bytes);
-        blob.extend_from_slice(&(st.history.len() as u32).to_le_bytes());
-        for t in &st.history {
-            blob.extend_from_slice(&t.to_le_bytes());
-        }
-        let off = blob.len();
-        blob.resize(off + conv_bytes, 0);
-        gpu.copy_d2h_on_stream(st.conv, &mut blob[off..], stream)?;
-        Ok(blob)
-    }
-
-    /// Restore the blob from [`Self::snapshot_aux`] on a prefix-cache hit.
-    pub fn restore_aux(&self, blob: &[u8], gpu: &dyn GpuBackend, stream: u64) -> Result<()> {
-        let mut st = self
-            .state
-            .lock()
-            .map_err(|_| anyhow::anyhow!("PLE state mutex poisoned"))?;
-        anyhow::ensure!(blob.len() >= 4, "PLE aux blob truncated");
-        let n = u32::from_le_bytes(blob[..4].try_into().unwrap()) as usize;
-        let conv_bytes = self.state_len * self.hc_mult * self.hidden * 4;
-        anyhow::ensure!(
-            blob.len() == 4 + n * 4 + conv_bytes,
-            "PLE aux blob size mismatch"
-        );
-        st.history = blob[4..4 + n * 4]
-            .chunks_exact(4)
-            .map(|c| u32::from_le_bytes(c.try_into().unwrap()))
-            .collect();
-        st.prestaged_va = None;
-        gpu.copy_h2d_async(&blob[4 + n * 4..], st.conv, stream)?;
-        Ok(())
-    }
+    // Marconi aux-state (snapshot_aux / restore_aux) moved to
+    // `aux_state.rs` (≤500 LoC split).
 
     /// Restore the prestaged state after a failed CUDA-graph capture attempt
-    /// consumed it: `slots_dev` still holds this step's slots and history has
-    /// already advanced, so the eager re-run only needs the VA back. No-op if
-    /// nothing was ever staged (idempotent if the forward never consumed it).
+    /// (the eager replay re-runs `forward`, which consumed `prestaged_va`).
     pub fn rearm(&self) {
         if let Ok(mut st) = self.state.lock()
             && st.last_staged_va != 0
@@ -522,3 +481,9 @@ pub struct PleWeights {
     pub norm_conv: DenseWeight,
     pub conv1d: DenseWeight,
 }
+
+// Child module (not sibling): the aux fns read PleLayer private
+// fields, and only a CHILD module sees them. Same #[path] trick
+// qsa.rs uses for its tests.
+#[path = "aux_state.rs"]
+mod aux_state;
