+++ crates/spark-model/src/weight_loader/qwen4_exp.rs
@@ -52,6 +52,7 @@ use crate::layer::TransformerLayer;
 use crate::weight_loader::ModelWeightLoader;
 use crate::weight_map::{DenseWeight, MtpWeights, dense};
 
+mod aux;
 mod ffn;
 mod hc;
 mod ple;
@@ -313,51 +314,9 @@ impl ModelWeightLoader for Qwen4ExpWeightLoader {
             let mut layer = layer;
             if config.hc_mult > 0 {
                 let (attn, ffn) = hc::load_layer_sites(store, &lp, config)?;
-                attach_hc(&mut layer, i, attn, ffn, hc_head.clone(), config)?;
+                aux::attach_hc(&mut layer, i, attn, ffn, hc_head.clone(), config)?;
             }
-            // QSA indexer on the 12 full-attention layers (#753 phase G).
-            // Decode-side selection; inert below the budget by arithmetic.
-            // ATLAS_QSA_DISABLE=1 skips the attach for A/B — decode then runs
-            // DENSE past the budget, which is NOT the reference model.
-            if config.index_topk > 0
-                && std::env::var("ATLAS_QSA_DISABLE").as_deref() != Ok("1")
-                && store.contains(&format!("{lp}.self_attn.indexer.index_qk_proj.weight"))
-            {
-                // Already device-resident in the store; the indexer holds the
-                // pointers (the store keeps the weights alive for the model's
-                // lifetime, same as every other layer weight).
-                let up = |name: &str| -> Result<spark_runtime::gpu::DevicePtr> {
-                    Ok(store
-                        .get(&format!("{lp}.self_attn.indexer.{name}"))
-                        .with_context(|| format!("qwen4_exp layer {i}: indexer {name}"))?
-                        .ptr)
-                };
-                let qsa = crate::layers::qsa::QsaIndexer::new(
-                    up("index_qk_proj.weight")?,
-                    up("q_layernorm.weight")?,
-                    up("k_layernorm.weight")?,
-                    config.index_n_heads,
-                    config.index_head_dim,
-                    config.index_compress_ratio,
-                    config.index_topk,
-                    (config.head_dim as f64 * config.partial_rotary_factor) as usize,
-                    config.rope_theta as f32,
-                    config.rms_norm_eps as f32,
-                    config.hidden_size,
-                    config.num_key_value_heads,
-                    config.head_dim,
-                    gpu,
-                )
-                .with_context(|| format!("qwen4_exp layer {i}: QSA indexer"))?;
-                let any = layer
-                    .as_any_mut()
-                    .ok_or_else(|| anyhow::anyhow!("qwen4_exp layer {i}: no as_any_mut for QSA"))?;
-                any.downcast_mut::<crate::layers::Qwen3AttentionLayer>()
-                    .ok_or_else(|| {
-                        anyhow::anyhow!("qwen4_exp layer {i}: indexer on a non-attention layer")
-                    })?
-                    .set_qsa(qsa);
-            }
+            aux::attach_qsa(&mut layer, i, &lp, store, config, gpu)?;
             // PLE lands on exactly one layer, which on this checkpoint is a
             // GDN one. `load` returns None for every other layer.
             let ple_layer = if ple_off {
@@ -366,17 +325,7 @@ impl ModelWeightLoader for Qwen4ExpWeightLoader {
                 ple::load(store, config, i, max_ple_tokens, gpu)?
             };
             if let Some(p) = ple_layer {
-                let any = layer
-                    .as_any_mut()
-                    .ok_or_else(|| anyhow::anyhow!("qwen4_exp layer {i}: no as_any_mut for PLE"))?;
-                let l = any
-                    .downcast_mut::<crate::layers::Qwen3SsmLayer>()
-                    .ok_or_else(|| {
-                        anyhow::anyhow!(
-                            "qwen4_exp layer {i} carries PLE but is not a GDN layer;                              the injection has nowhere to go"
-                        )
-                    })?;
-                l.set_ple(p);
+                aux::attach_ple(&mut layer, i, p)?;
             }
             layers.push(layer);
             hc_bytes += f2.saturating_sub(free_now(gpu));
@@ -453,21 +402,7 @@ impl ModelWeightLoader for Qwen4ExpWeightLoader {
         config: &ModelConfig,
         gpu: &dyn GpuBackend,
     ) -> Result<DenseWeight> {
-        let mixer = mixer_prefix(config);
-        anyhow::ensure!(
-            store.contains(&format!("{mixer}.hc_norm.weight")),
-            "qwen4_exp: no `{mixer}.hc_norm.weight` — this architecture is \
-             supposed to keep its final normalization in the hyper-connection \
-             mixer, and it is not there. Refusing rather than guessing."
-        );
-        tracing::warn!(
-            "qwen4_exp: final norm is a PLACEHOLDER. The real one is \
-             `{mixer}.hc_norm` [{}], applied while collapsing the {} residual \
-             streams — that is mHC work, not a final-norm substitution.",
-            config.hc_mult * config.hidden_size,
-            config.hc_mult,
-        );
-        ones_norm(config.hidden_size, gpu)
+        aux::final_norm_placeholder(store, config, gpu)
     }
 
     fn load_lm_head(
@@ -539,59 +474,4 @@ fn embed_prefix(config: &ModelConfig) -> String {
 /// The model-level hyper-connection mixer that collapses the residual streams.
 fn mixer_prefix(config: &ModelConfig) -> String {
     format!("{}.hyper_connection_mixer", embed_prefix(config))
-}
-
-/// Attach both hyper-connection sites to a freshly built layer.
-///
-/// `set_hc_weights` lives on `Qwen3AttentionLayer`, but `load_layers` hands
-/// back `Box<dyn TransformerLayer>`, so the concrete type has to be recovered.
-/// A failure here is a hard error rather than a skip: a layer that silently
-/// keeps no mHC weights would run attention on an unmixed stream and produce
-/// plausible, wrong activations.
-fn attach_hc(
-    layer: &mut Box<dyn TransformerLayer>,
-    idx: usize,
-    attn: crate::layers::qwen3_attention::HcSiteWeights,
-    ffn: crate::layers::qwen3_attention::HcSiteWeights,
-    head: Option<crate::layers::qwen3_attention::HcHeadWeights>,
-    config: &ModelConfig,
-) -> Result<()> {
-    use crate::layers::qwen3_attention::HcWeights;
-    // Hard error, never a skip. A layer that quietly kept no mHC weights
-    // would run attention on a stream it never mixed and emit plausible,
-    // wrong activations — with nothing in the log.
-    let any = layer.as_any_mut().ok_or_else(|| {
-        anyhow::anyhow!("qwen4_exp layer {idx}: no as_any_mut, cannot attach mHC weights")
-    })?;
-    // TWO concrete layer types carry mHC here: the 12 full-attention layers
-    // are `Qwen3AttentionLayer`, the 36 GDN layers are `Qwen3SsmLayer`.
-    // DeepSeek-V4 only ever needed the first, which is why the second had to
-    // learn `set_hc_weights`.
-    let w = HcWeights {
-        attn,
-        ffn,
-        head,
-        hc_mult: config.hc_mult,
-        sinkhorn_iters: 0,
-        hc_eps: config.rms_norm_eps as f32,
-        // MODEL layer indices, not attention-layer ones. With a 3:1
-        // GDN:attention interleave, model layer 0 is GDN and the last model
-        // layer (47) is attention — so the layer that seeds the highway and
-        // the layer that collapses it are DIFFERENT concrete types, and
-        // neither is identified by `attn_layer_idx`.
-        is_first_model_layer: idx == 0,
-        is_last_model_layer: idx + 1 == config.num_hidden_layers,
-    };
-    if let Some(l) = any.downcast_mut::<crate::layers::Qwen3AttentionLayer>() {
-        l.set_hc_weights(w);
-        return Ok(());
-    }
-    if let Some(l) = any.downcast_mut::<crate::layers::Qwen3SsmLayer>() {
-        l.set_hc_weights(w);
-        return Ok(());
-    }
-    anyhow::bail!(
-        "qwen4_exp layer {idx}: mHC weights have nowhere to go — the layer is \
-         neither Qwen3AttentionLayer nor Qwen3SsmLayer"
-    )
 }
